+++ tls/src/rustls.rs
@@ -7,7 +7,7 @@ use core::{
 
 use std::io;
 
-pub use rustls::*;
+pub use rustls_crate::*;
 
 use xitca_io::io::{AsyncIo, Interest, Ready};