+++ http/src/h3/body.rs
@@ -1,20 +1,28 @@
-use std::{
+use core::{
     pin::Pin,
     task::{Context, Poll},
 };
 
-use futures_core::stream::{BoxStream, Stream};
+use ::h3::server::RequestStream;
+use futures_core::stream::Stream;
+use h3_quinn::RecvStream;
 
-use crate::{bytes::Bytes, error::BodyError};
+use crate::{
+    bytes::{Buf, Bytes},
+    error::BodyError,
+};
 
 /// Request body type for Http/3 specifically.
-pub struct RequestBody(pub(super) BoxStream<'static, Result<Bytes, h3::Error>>);
+pub struct RequestBody(pub(super) RequestStream<RecvStream, Bytes>);
 
 impl Stream for RequestBody {
     type Item = Result<Bytes, BodyError>;
 
     fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        self.get_mut().0.as_mut().poll_next(cx).map_err(Into::into)
+        self.get_mut()
+            .0
+            .poll_recv_data(cx)?
+            .map(|res| res.map(|buf| Ok(Bytes::copy_from_slice(buf.chunk()))))
     }
 }